Add swagger documentation for rides api v1 (#1)
The PR adds a swagger documentation for the Rides API, with a new dependency to swagger-ui-express. Web documentation can be viewed with /api-documentation/v1. Upon API deployment, the documentation should also be available. The swagger file is prefixed with api-v as to indicate the API version it is documenting, since it is possible to have multiple documentation support multiple versions of the API.